+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * support for determining whether PDF downloads are georeferenced
 * additional date types (creation, revision, released)
 * support for custom media types (GeoPDF, zipped ShapeFile etc.)
+* initial record export component
 
 ### Changed